refactor frequency modulation

+++ src/engine.h
@@ -60,7 +60,7 @@ public:
 	void					midi_in(uint8_t c, uint8_t d1, uint8_t d2);
 
 public:
-	uint8_t					update(int32_t* samples, size_t n);
+	uint32_t				update(int32_t* samples, size_t n);
 
 public:
 							SynthEngine();
